+++ tests/fscripts/demo_str2.fs
@@ -1,0 +1,3 @@
+a = prompt('input a number:');
+print(int(a)+100);
+